+++ NEWS.md
@@ -1,3 +1,5 @@
 # namap 0.0.0.9000
 
 - Original commit
+- Added function `map_names()` including unit test
+- Update `README.Rmd`